+++ air/src/execution_step/air/call/call_result_setter.rs
@@ -22,7 +22,7 @@ use crate::UncatchableError;
 
 use air_interpreter_data::CallResult;
 use air_interpreter_data::TracePos;
-use air_interpreter_data::Value;
+use air_interpreter_data::ValueRef;
 use air_parser::ast::CallOutputValue;
 use air_trace_handler::merger::ValueSource;
 use air_trace_handler::TraceHandler;
@@ -32,11 +32,15 @@ pub(crate) fn populate_context_from_peer_service_result<'i>(
     output: &CallOutputValue<'i>,
     exec_ctx: &mut ExecutionCtx<'i>,
 ) -> ExecutionResult<CallResult> {
-    let result_value = executed_result.result.clone();
+    let cid = exec_ctx
+        .cid_tracker
+        .record_value(executed_result.result.clone())
+        .map_err(UncatchableError::from)?;
+
     match output {
         CallOutputValue::Scalar(scalar) => {
             exec_ctx.scalars.set_scalar_value(scalar.name, executed_result)?;
-            Ok(CallResult::executed_scalar(result_value))
+            Ok(CallResult::executed_scalar(cid))
         }
         CallOutputValue::Stream(stream) => {
             let value_descriptor = StreamValueDescriptor::new(
@@ -47,30 +51,36 @@ pub(crate) fn populate_context_from_peer_service_result<'i>(
                 stream.position,
             );
             let generation = exec_ctx.streams.add_stream_value(value_descriptor)?;
-            Ok(CallResult::executed_stream(result_value, generation))
+            Ok(CallResult::executed_stream(cid, generation))
         }
         // by the internal conventions if call has no output value,
         // corresponding data should have scalar type
-        CallOutputValue::None => Ok(CallResult::executed_scalar(result_value)),
+        CallOutputValue::None => Ok(CallResult::executed_scalar(cid)),
     }
 }
 
 pub(crate) fn populate_context_from_data<'i>(
-    value: Value,
+    value: ValueRef,
     tetraplet: RcSecurityTetraplet,
     trace_pos: TracePos,
     value_source: ValueSource,
     output: &CallOutputValue<'i>,
     exec_ctx: &mut ExecutionCtx<'i>,
-) -> ExecutionResult<Value> {
+) -> ExecutionResult<ValueRef> {
     match (output, value) {
-        (CallOutputValue::Scalar(scalar), Value::Scalar(value)) => {
-            let result = ValueAggregate::new(value.clone(), tetraplet, trace_pos);
+        (CallOutputValue::Scalar(scalar), ValueRef::Scalar(cid)) => {
+            let value = exec_ctx
+                .get_value_by_cid(&cid)
+                .ok_or_else(|| UncatchableError::ValueForCidNotFound(cid.clone()))?;
+            let result = ValueAggregate::new(value, tetraplet, trace_pos);
             exec_ctx.scalars.set_scalar_value(scalar.name, result)?;
-            Ok(Value::Scalar(value))
+            Ok(ValueRef::Scalar(cid))
         }
-        (CallOutputValue::Stream(stream), Value::Stream { value, generation }) => {
-            let result = ValueAggregate::new(value.clone(), tetraplet, trace_pos);
+        (CallOutputValue::Stream(stream), ValueRef::Stream { cid, generation }) => {
+            let value = exec_ctx
+                .get_value_by_cid(&cid)
+                .ok_or_else(|| UncatchableError::ValueForCidNotFound(cid.clone()))?;
+            let result = ValueAggregate::new(value, tetraplet, trace_pos);
             let value_descriptor = StreamValueDescriptor::new(
                 result,
                 stream.name,
@@ -80,15 +90,15 @@ pub(crate) fn populate_context_from_data<'i>(
             );
             let resulted_generation = exec_ctx.streams.add_stream_value(value_descriptor)?;
 
-            let result = Value::Stream {
-                value,
+            let result = ValueRef::Stream {
+                cid,
                 generation: resulted_generation,
             };
             Ok(result)
         }
         // by the internal conventions if call has no output value,
         // corresponding data should have scalar type
-        (CallOutputValue::None, value @ Value::Scalar(_)) => Ok(value),
+        (CallOutputValue::None, value @ ValueRef::Scalar(_)) => Ok(value),
         (_, value) => Err(ExecutionError::Uncatchable(
             UncatchableError::CallResultNotCorrespondToInstr(value),
         )),